Add files via upload
Attract-Arcade-Horizontal.txt - Renamed using underscores for consistency
Attract_Arcade_off.sh - Added cleanup for daemon
Attract_Arcade_on.sh - Reworked loading of MRAs - any problems are with the MRAs not the script based on additional testing
  Added "next" command to daemon allowing you to skip to another MRA - usable via SSH for debugging
  Added output from daemon when loading MRA - visible via SSH for debugging

+++ Attract_Arcade_off.sh
@@ -1,14 +1,35 @@
+#!/bin/bash
+
+killall -q -9 S98attract
+rm -f /var/run/attract.pid
+
+# Verify the daemon exists
+if [ ! -f /etc/init.d/S98attract ]; then
+	echo "Attract mode is not present!"
+	exit 0
+fi
+
+# Disable the daemon
 mount | grep -q "on / .*[(,]ro[,$]" && RO_ROOT="true"
 [ "${RO_ROOT}" == "true" ] && mount / -o remount,rw
 mv /etc/init.d/S98attract /etc/init.d/_S98attract > /dev/null 2>&1
 sync
 [ "${RO_ROOT}" == "true" ] && mount / -o remount,ro
-sed -i "s/.*bootcore=.*/#bootcore=/" /media/fat/MiSTer.ini
 
+# Legacy - remove bootcore
+#sed -i "s/.*bootcore=.*/#bootcore=/" /media/fat/MiSTer.ini
+
+# Reboot
 echo "Attract mode is off and"
 echo "inactive at startup."
-echo "Done!"
-echo "Rebooting in 5s"
+echo ""
+echo "Rebooting in..."
 sync
-sleep 5 && fpga /media/fat/menu.rbf
+for i in {5..1}; do
+	echo "${i} seconds"
+	sleep 1
+done
+
+echo "load_core /media/fat/menu.rbf" > /dev/MiSTer_cmd
+
 exit 0
